Fixes for warnings and compile failure under VC++
git-svn-id: svn://sporkhack.com/nh@185 0b9a868d-f130-0410-bc2a-adc7ae8015a6

+++ src/worn.c
@@ -351,11 +351,11 @@ register struct monst *mon;
 
 	/* Tweak the monster's AC a bit according to its level */
 	div = (mon->m_lev > 20) ? 6 : 5;
-	bonus = mon->m_lev/2;
+	bonus = (float)(mon->m_lev/2);
 	bonus *= bonus/div;
 	if (bonus > 16) { bonus = 16; }
 	if (bonus < 0) { bonus = 0; }
-	base -= bonus;
+	base -= (int)bonus;
 
 	return base;
 }
